aggregate serialization and deserialization logic into own module
